50 fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -2,18 +2,18 @@
 int main(void)
 {
 	int n;
-	unsigned int firstTerm = 1;
-	unsigned int secondTerm = 2;
+	unsigned long int firstTerm = 1;
+	unsigned long int secondTerm = 2;
 	int nextTerm;
 	int i;
 	n = 50;
-	printf("%u, %u, ", firstTerm, secondTerm);
+	printf("%lu, %lu, ", firstTerm, secondTerm);
 	for (i = 2; i < n; i++)
 	{
 		nextTerm = firstTerm + secondTerm;
 		firstTerm = secondTerm;
 		secondTerm = nextTerm;
-		printf("%u",secondTerm);
+		printf("%lu",secondTerm);
 		if(i < n - 1)
 		{
 			printf(", ");
